rename: 기존(old) 도메인 모델을 application/domain 하위로 이동
