+++ src/main/java/com/ddudu/application/domain/goal/domain/enums/GoalStatus.java
@@ -1,0 +1,7 @@
+package com.ddudu.application.domain.goal.domain.enums;
+
+
+public enum GoalStatus {
+  IN_PROGRESS,
+  DONE
+}
