--- src/main/java/com/ddudu/old/goal/domain/GoalStatus.java
@@ -1,7 +1,0 @@
-package com.ddudu.old.goal.domain;
-
-
-public enum GoalStatus {
-  IN_PROGRESS,
-  DONE
-}
