--- src/main/java/com/ddudu/old/goal/domain/PrivacyType.java
@@ -1,7 +1,0 @@
-package com.ddudu.old.goal.domain;
-
-public enum PrivacyType {
-  PRIVATE,
-  FOLLOWER,
-  PUBLIC
-}
