+++ src/main/java/com/ddudu/old/persistence/dao/todo/TodoDaoCustom.java
@@ -1,7 +1,7 @@
 package com.ddudu.old.persistence.dao.todo;
 
+import com.ddudu.application.domain.goal.domain.enums.PrivacyType;
 import com.ddudu.infrastructure.persistence.entity.UserEntity;
-import com.ddudu.old.goal.domain.PrivacyType;
 import com.ddudu.old.persistence.entity.TodoEntity;
 import com.ddudu.old.todo.dto.response.TodoCompletionResponse;
 import java.time.LocalDateTime;
