+++ src/main/java/com/ddudu/old/persistence/dao/todo/TodoDaoImpl.java
@@ -2,8 +2,8 @@ package com.ddudu.old.persistence.dao.todo;
 
 import static com.ddudu.old.persistence.entity.QTodoEntity.todoEntity;
 
+import com.ddudu.application.domain.goal.domain.enums.PrivacyType;
 import com.ddudu.infrastructure.persistence.entity.UserEntity;
-import com.ddudu.old.goal.domain.PrivacyType;
 import com.ddudu.old.persistence.entity.TodoEntity;
 import com.ddudu.old.todo.domain.TodoStatus;
 import com.ddudu.old.todo.dto.response.TodoCompletionResponse;
