+++ src/main/java/com/ddudu/old/todo/domain/Todo.java
@@ -5,7 +5,7 @@ import static java.util.Objects.isNull;
 
 import com.ddudu.application.domain.user.domain.User;
 import com.ddudu.old.common.domain.BaseDomain;
-import com.ddudu.old.goal.domain.Goal;
+import com.ddudu.application.domain.goal.domain.Goal;
 import com.ddudu.old.todo.exception.TodoErrorCode;
 import com.ddudu.presentation.api.exception.InvalidParameterException;
 import java.time.LocalDateTime;
