+++ src/main/java/com/ddudu/old/todo/domain/TodoRepository.java
@@ -1,7 +1,7 @@
 package com.ddudu.old.todo.domain;
 
+import com.ddudu.application.domain.goal.domain.enums.PrivacyType;
 import com.ddudu.application.domain.user.domain.User;
-import com.ddudu.old.goal.domain.PrivacyType;
 import com.ddudu.old.todo.dto.response.TodoCompletionResponse;
 import java.time.LocalDateTime;
 import java.util.List;
