+++ src/main/java/com/ddudu/old/todo/dto/response/TodoResponse.java
@@ -1,6 +1,6 @@
 package com.ddudu.old.todo.dto.response;
 
-import com.ddudu.old.goal.domain.Goal;
+import com.ddudu.application.domain.goal.domain.Goal;
 import com.ddudu.old.todo.domain.Todo;
 import lombok.Builder;
